Fix leadearboard showing on load
This fixes the situation where the empty leaderboard flashses on the screen during loading if no players are loaded yet.

this.players and this.guesses are not JavaScript objects actually holding players but an Observer object that facilitates the "reactive" behaviour of the app. Thus, they are always non-empty. Checking explicitly for "keys" fixes the problem, and leaderboard show only when actually all players have guessed.