Maybe fix the evil overlay
I am not completely sure why this helps, nor why it used v-show instead of value before but it seems to help - at least it stops reproducing on dev server.

My guess is that the previous version triggered some kind of race condition, potentially not in our code but in the framework, but :value is evaluated differently.