Add helper functions to map_util.js, slight interface fix
Details:
getClosestPanorama now takes an object with {lat, lng} fields,
and returns and object with {lat, lng} fields, rather than
{lat(), lng()} methods. This helps unify the (currently)
many ways in which lat-longs are passed.

Adds a jumpByDistanceAndBearing method, which takes a `point`,
a `distance_km` and a `bearing_deg`, and tries to jump to a panorama
at most `distance_km` away from `point` in the direction `bearing_deg`.
We try to find a point exactly `distance_km` away, but if that fails,
we make a total of 10 attempts to jump to any point in the direction
using a linear distance backoff.

Adds a centerPoint method, which takes an array of points and computes
the mathematical mean of the lat and lng of these points.
This results in a crude centerpoint.

The last two methods aren't used yet, but will be used in a followup
CL.