Add optional jump buttons to Street.vue component
Additionally, changes how the component communicates
with parent components. Before it was assumed that the
parent component will populate the position with the
mapPosition prop, and there was never a need to communicate
the true streetview position back to the parent.

Now the parent sets the initial position with the
initialMapPosition prop, and afterwards may subscribe to
changes to the streetview location through the position_changed
event.