Rendezvous game state generation and lobby setting
This change creates a lobby option for picking between
the available game modes (the option is hidden, since the main
rendezvous logic isn't implemented yet).
The change also initializes the rendezvous game state.
The game mode and list of players are now both stored in the game
generation Store. The game mode is set in the LobbyOptions,
the list of players in the Lobby.
A new game generation is triggered when either of these changes.