Add an experiment store
The idea behind this change is to allow one to gate
experimental features behind a cookie-based flag.
The flag can easily be set or unset in the vue devtools
(though the method is a bit obtuse).
To do this:
1. Open the Vue devtools plugin.
2. Find the Header component.
3. Click it, it should become green and a "= $vm0" should show up
next to it.
4. In the console type $vm0.setRendezvousEnabled(true);
5. Refresh the page.
Rendezvous should now be persistently enabled (even after page
refreshes, since the experiment is stored in cookies). The header
will change its color and text to make this more visible.
To disable the experiment, just use the same process as above,
only changing the call to set to false.

Currently for rendezvous only the host needs to have the experiment
enabled.