Lint, don't use undefined MarkerMap prop, fix error handling in Lobby
The MarkerMap prop in question will be added in a future change,
using it in the previous change was an oversight.
Also, it turns out we were using function() to define the callback
for confirmAction, but the problem is that doesn't carry over
the correct "this" for cleanUpAndChangeView to work.
As a result none of the confirmation buttons for errors in Lobby
worked.
I only fixed this issue in Lobby, though I suspect it's more widespread.